requirements.txt
pytesseract~=0.3.7
Flask~=1.1.2
Pillow~=8.2.0

scikit-image~=0.18.1
deskew~=0.10.26
numpy~=1.20.2